requirements.txt
# Core Data Processing
pyspark==3.4.1
pandas==2.0.3
numpy==1.24.3

# Database
pymongo==4.5.0
psycopg2-binary==2.9.7

# Streaming
kafka-python==2.0.2
confluent-kafka==2.1.1

# Web Framework and Dashboard
flask==2.3.2
flask-socketio==5.3.4
plotly==5.15.0
dash==2.11.1
dash-bootstrap-components==1.4.1

# Machine Learning
scikit-learn==1.3.0
tensorflow==2.13.0
xgboost==1.7.6

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2

# HTTP Requests and Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2

# Configuration and Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
schedule==1.2.0

# Testing
pytest==7.4.0
pytest-mock==3.11.1

# Development
jupyter==1.0.0
ipykernel==6.25.0 